refactor: move gap impls to `smith_waterman::algo`
